Newly Organized App.
 public/app - works, ...public doesn`t work???
