+++ config.js
@@ -1,0 +1,5 @@
+module.exports = {
+    'port' : process.env.PORT || 8080,
+    'database' : 'mongodb://user:kikiriki07@ds111798.mlab.com:11798/contact_list',
+    'secret' : 'ilovescotchscotchyscotchscotch'
+};
